Default CI Method

For certain resampling methods, there are default confidence interval methods. See mlr3::mlr_reflections$default_ci_methods for a selection. This measure will select the appropriate CI method depending on the class of the used Resampling.

Parameters

Only those from MeasureAbstractCi.

Super classes

mlr3::Measure -> mlr3inferr::MeasureAbstractCi -> Measure

Methods

Public methods

MeasureCi$new()

MeasureCi$aggregate()

MeasureCi$clone()

Inherited methods

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureCi$new(measure)

Arguments

measure

(Measure or character(1))
A measure of ID of a measure.

Method aggregate()

Obtain a point estimate, as well as lower and upper CI boundary.

Usage

MeasureCi$aggregate(rr)

Arguments

rr

(ResampleResult)
Resample result.

Returns

named numeric(3)

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureCi$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

rr = resample(tsk("sonar"), lrn("classif.featureless"), rsmp("holdout"))
rr$aggregate(msr("ci", "classif.acc"))
#>       classif.acc classif.acc.lower classif.acc.upper 
#>         0.3913043         0.2753062         0.5073025 
# is the same as:
rr$aggregate(msr("ci.holdout", "classif.acc"))
#>       classif.acc classif.acc.lower classif.acc.upper 
#>         0.3913043         0.2753062         0.5073025